requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
google-generativeai==0.3.2
chromadb==0.4.22
python-dotenv==1.0.0
pillow>=10.0.0
pydantic>=2.0.0
python-multipart==0.0.6
aiofiles==23.2.1
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0
